Refactor: store 역할 별 분리 (상태관리/타입관리/비즈니스 로직)
