+++ src/components/todo/TodoList.tsx
@@ -1,5 +1,5 @@
 import { Todo } from '@/types/todo';
-import { useTodoStore } from '@/store/todoStore';
+import { useTodoStore } from '@/store/todo/todoStore';
 import TodoItem from './TodoItem';
 
 interface TodoListProps {
